+++ apps/forms/scaffolds/blocks-editor/blocks/divider-block.tsx
@@ -1,0 +1,54 @@
+"use client";
+
+import {
+  DividerHorizontalIcon,
+  DotsHorizontalIcon,
+  TrashIcon,
+} from "@radix-ui/react-icons";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@editor-ui/dropdown-menu";
+import { EditorFlatFormBlock } from "@/scaffolds/editor/state";
+import { BlockHeader, FlatBlockBase, useDeleteBlock } from "./base-block";
+
+export function DividerBlock({ id }: EditorFlatFormBlock) {
+  const deleteBlock = useDeleteBlock();
+
+  return (
+    <FlatBlockBase>
+      <BlockHeader>
+        <div className="flex flex-row items-center gap-8">
+          <div className="flex flex-col gap-1">
+            <span className="flex flex-row gap-2 items-center">
+              <DividerHorizontalIcon />
+              Divider
+            </span>
+          </div>
+        </div>
+        <div>
+          <DropdownMenu modal={false}>
+            <DropdownMenuTrigger asChild>
+              <button>
+                <DotsHorizontalIcon />
+              </button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent>
+              <DropdownMenuItem onClick={() => deleteBlock(id)}>
+                <TrashIcon />
+                Delete Block
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        </div>
+      </BlockHeader>
+      <div>
+        <div className="bg-neutral-200 rounded overflow-hidden border border-black/20 aspect-auto">
+          <hr />
+        </div>
+      </div>
+    </FlatBlockBase>
+  );
+}
